Update models.xsl
Bug with dynamic-array inside hiding filter :
When switching the filter from show to hide, it empties everything inside the filter.
This includes the dynamic-array's counter.

But when the counter is empty, it causes a bug in xforms-ready, which doesn't show the pop-up "welcome-back" and show the "welcome" page with the navigation buttons of the current page (no button when last page with the proof of filing link)