[CI/CD] version variable
Change hard-coded version workflow to a dynamic variable, using `printVersion` plugin defined in the build.

The Docker image produced contains a jar named `eno-ws.jar` (without version in its name so that the `Dockerfile` does not have to be modified when version changes), but is published in Docker Hub with the current version's tag.

Other details:

- use wrapper in workflow
- upgrade version of publish docker action
- effectively remove version in Dockerfile
- changed relative paths in workflow & docker image